pyproject.toml
[tool.poetry]
name = "trackcoin"
version = "1.0.0"
description = "A bot made for use in patrimonial increase, free, easy, game changer for everyone."
authors = ["venysssssssssss <shablaur7@hotmail.com>"]
license = "APACHE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.44.2"
datasets = "^3.0.0"
faiss-cpu = "^1.8.0.post1"
torch = "^2.4.1"
torchvision = "^0.19.1"
torchaudio = "^2.4.1"
sentencepiece = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
